api/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pandas>=2.0.0
numpy>=1.20.0
scikit-learn>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
python-dotenv>=1.0.0
pydantic>=2.0.0 